rss: Extract bbcode module
